+++ lib/dbee/providers/active_record_provider/makers/order.rb
@@ -10,17 +10,10 @@
 module Dbee
   module Providers
     class ActiveRecordProvider
-      class ExpressionBuilder
+      module Makers # :nodoc: all
         # Derives Arel#order predicates.
-        class OrderMaker
+        class Order
           include Singleton
-
-          SORTER_EVALUATORS = {
-            Query::Sorters::Ascending => ->(column) { column },
-            Query::Sorters::Descending => ->(column) { column.desc }
-          }.freeze
-
-          private_constant :SORTER_EVALUATORS
 
           def make(sorter, arel_column)
             method = SORTER_EVALUATORS[sorter.class]
@@ -29,6 +22,13 @@ module Dbee
 
             method.call(arel_column)
           end
+
+          SORTER_EVALUATORS = {
+            Query::Sorters::Ascending => ->(column) { column },
+            Query::Sorters::Descending => ->(column) { column.desc }
+          }.freeze
+
+          private_constant :SORTER_EVALUATORS
         end
       end
     end